refactor: organize folder structure for services and data models
