+++ src/app/services/api/auth-service.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { AuthService } from './auth-service';
+import { AuthService } from '../api/auth-service';
 
 describe('AuthService', () => {
   let service: AuthService;
