+++ src/app/services/guards/auth-guard.ts
@@ -1,6 +1,6 @@
 import { inject } from '@angular/core';
 import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../auth-service';
+import { AuthService } from '../api/auth-service';
 
 export const authGuard: CanActivateFn = (route, state) => {
   const authService = inject(AuthService);
