+++ src/app/services/interceptors/auth-interceptor.ts
@@ -1,6 +1,6 @@
 import { HttpInterceptorFn } from '@angular/common/http';
 import { inject } from '@angular/core';
-import { AuthService } from '../auth-service';
+import { AuthService } from '../api/auth-service';
 import { Router } from '@angular/router';
 import { catchError, throwError } from 'rxjs';
