+++ src/app/services/ui/sidebar-service.ts
@@ -1,5 +1,5 @@
 import { computed, inject, Injectable, signal } from '@angular/core';
-import { AuthService } from '../auth-service';
+import { AuthService } from '../api/auth-service';
 
 @Injectable({
   providedIn: 'root'
